消費側が Go と editor-ui だけなので、OpenAPI に固定幅を宣言する契約上の意義が薄い。tsp の符号付き整数を
`int32` から `integer`(→ Go `int`) に統一し、ドメインの `int`
との境界に出ていた変換キャストを解消する。#619 で balance に適用した方針の全体展開。判断基準は `ruins-tsp` skill
に明記済み。

## 変更
- `types.tsp`/`main.tsp`/`save.tsp`/`resp.tsp` の `int32`（スカラー + インライン）を
`integer` に。生成 Go 型が `int` になる。
- `resolve.go` の深度比較を `int` 直接比較に修正（唯一の型不一致ビルドエラー）。
- 冗長化した変換キャスト **53件**を `unconvert -apply` で自動撤去（`raw.go` 中心に
profession/disassembly/disassemble/balance）。

## 据え置き
- `uint8`(ColorChannel)・`uint32`(EntityIndex/Generation)
は符号なし・幅に意味があり、serde と ECS のエンティティ ID が依存するため据え置く。

## 影響範囲
- editor-ui の生成 TS は `int32`/`integer` とも `number` で **不変**。
- 直接のビルドエラーは resolve.go の4件のみ。残りは冗長キャストの機械的除去。

## 検証
- Go `make check` 全通過（0 issues・脆弱性なし）。

+++ internal/mapplanner/resolve.go
@@ -21,7 +21,7 @@ func resolveEnemyEntries(rawMaster *oapi.Raws, tableName string, depth int) ([]S
 	}
 	result := make([]SpawnEntry, 0, len(enemyTable.Entries))
 	for _, entry := range enemyTable.Entries {
-		if int32(depth) < entry.MinDepth || int32(depth) > entry.MaxDepth {
+		if depth < entry.MinDepth || depth > entry.MaxDepth {
 			continue
 		}
 		pack, err := consts.ParseDice(entry.Pack)
@@ -52,7 +52,7 @@ func resolveItemSources(rawMaster *oapi.Raws, tableName string, depth int) ([]it
 	}
 	result := make([]itemGroupRef, 0, len(itemTable.Entries))
 	for _, entry := range itemTable.Entries {
-		if int32(depth) < entry.MinDepth || int32(depth) > entry.MaxDepth {
+		if depth < entry.MinDepth || depth > entry.MaxDepth {
 			continue
 		}
 		result = append(result, itemGroupRef{GroupID: entry.Id, Weight: entry.Weight})
